Re-implement changes clobbered by Mesa 6 merge. Build & warning fixes.

+++ extras/Mesa/src/mesa/drivers/dri/sis/sis_alloc.c
@@ -24,7 +24,7 @@ OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
 USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 **************************************************************************/
-/* $XFree86: xc/lib/GL/mesa/src/drv/sis/sis_alloc.c,v 1.7 2001/01/08 01:07:29 martin Exp $ */
+/* $XFree86: xc/extras/Mesa/src/mesa/drivers/dri/sis/sis_alloc.c,v 1.1.1.1tsi Exp $ */
 
 /*
  * Authors:
@@ -170,7 +170,7 @@ sisAllocZStencilBuffer( sisContextPtr smesa )
    memset( &smesa->zClearPacket, 0, sizeof(ENGPACKET) );
 
    smesa->zClearPacket.dwSrcPitch = (z_depth == 2) ? 0x80000000 : 0xf0000000;
-   smesa->zClearPacket.dwDestBaseAddr = (GLint)(addr -
+   smesa->zClearPacket.dwDestBaseAddr = (unsigned long)(addr -
       (unsigned long)smesa->FbBase);
    smesa->zClearPacket.wDestPitch = width2;
    smesa->zClearPacket.stdwDestPos.wY = 0;
@@ -218,7 +218,7 @@ sisAllocBackbuffer( sisContextPtr smesa )
    addr = (char *)ALIGNMENT( (unsigned long)addr, DRAW_BUFFER_HW_ALIGNMENT );
 
    smesa->backbuffer = addr;
-   smesa->backOffset = (GLint)(addr - (unsigned long)smesa->FbBase);
+   smesa->backOffset = (unsigned long)(addr - (unsigned long)smesa->FbBase);
    smesa->backPitch = width2 * depth;
 
    memset ( &smesa->cbClearPacket, 0, sizeof(ENGPACKET) );
